force production env when precompiling assets
I was having the same issue detailed at http://stackoverflow.com/questions/19086898/image-path-production-reference-is-not-working-in-rails-4

Assets in production were being served without finger-printed names.

Forcing the env to be production when compiling the assets locally fixed this - although with a catch - you need to have a 'production' database defined in database.yml  in order for it to work. I can just be the exact same as your development database - it'll never actually get used.

Signed-off-by: Joshua Paling <joshua.paling@gmail.com>